fix: fixed link in script
the link to fetch the ics file was this:

    /cal.ics

which worked under the localhost server but not in github pages because
it's an absolute link.

The request is interpreted as benraz.dev/cal.ics when it should really
be benraz.dev/wopl/cal.ics. So i've changed it to "cal.ics" instead to
make it relative to the current dir.

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>